+++ CHANGELOG.rst
@@ -3,7 +3,14 @@
 Changelog
 =========
 
+Version 0.2.0 - Unreleased
+--------------------------
+
+* The controller form save is now working with basic fields;
+
+
 Version 0.1.0 - Unreleased
 --------------------------
 
-Initial version.
+Initial version started to prototype application with starting models, form forge,
+dummy base controller form and some tests.